# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/restock/priorities": {
            "get": {
                "description": "Returns a paginated list of products that need restocking, sorted by urgency",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restock"
                ],
                "summary": "Get restock priorities",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.restockPrioritiesResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    }
                }
            }
        },
        "/stock": {
            "get": {
                "description": "Returns a paginated list of all product stocks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "List all product stocks",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.productStockResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new product stock entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Create a product stock",
                "parameters": [
                    {
                        "description": "Product stock data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.createProductStockRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.createResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    }
                }
            }
        },
        "/stock/category/{category}": {
            "get": {
                "description": "Returns a paginated list of product stocks filtered by category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Get product stocks by category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product category",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.productStockResponseDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    }
                }
            }
        },
        "/stock/{id}": {
            "get": {
                "description": "Returns a single product stock by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Get a product stock by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product stock ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.productStockResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    }
                }
            },
            "put": {
                "description": "Partially updates a product stock by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Update a product stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product stock ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateProductStockRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a product stock by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Delete a product stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product stock ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponseDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.createProductStockRequestDTO": {
            "type": "object",
            "properties": {
                "average_daily_sales": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "criticality_level": {
                    "type": "integer"
                },
                "current_stock": {
                    "type": "integer"
                },
                "lead_time_days": {
                    "type": "integer"
                },
                "minimum_stock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit_cost": {
                    "type": "number"
                }
            }
        },
        "http.createResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "uuid"
                }
            }
        },
        "http.errorResponseDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "http.productStockResponseDTO": {
            "type": "object",
            "properties": {
                "averageDailySales": {
                    "type": "integer",
                    "example": 10
                },
                "category": {
                    "type": "string",
                    "example": "engine"
                },
                "criticalityLevel": {
                    "type": "integer",
                    "example": 3
                },
                "currentStock": {
                    "type": "integer",
                    "example": 150
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "leadTimeDays": {
                    "type": "integer",
                    "example": 7
                },
                "minimumStock": {
                    "type": "integer",
                    "example": 50
                },
                "name": {
                    "type": "string",
                    "example": "Engine Oil Filter"
                },
                "unitCost": {
                    "type": "number",
                    "example": 25.5
                }
            }
        },
        "http.restockPrioritiesResponseDTO": {
            "type": "object",
            "properties": {
                "priorities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.restockPriorityResponseDTO"
                    }
                }
            }
        },
        "http.restockPriorityResponseDTO": {
            "type": "object",
            "properties": {
                "currentStock": {
                    "type": "integer",
                    "example": 15
                },
                "minimumStock": {
                    "type": "integer",
                    "example": 20
                },
                "name": {
                    "type": "string",
                    "example": "Engine Oil Filter"
                },
                "partId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "projectedStock": {
                    "type": "integer",
                    "example": -20
                },
                "urgencyScore": {
                    "type": "number",
                    "example": 45
                }
            }
        },
        "http.updateProductStockRequestDTO": {
            "type": "object",
            "properties": {
                "average_daily_sales": {
                    "type": "integer"
                },
                "criticality_level": {
                    "type": "integer"
                },
                "current_stock": {
                    "type": "integer"
                },
                "lead_time_days": {
                    "type": "integer"
                },
                "minimum_stock": {
                    "type": "integer"
                },
                "unit_cost": {
                    "type": "number"
                }
            }
        }
    }
}